DFT almost implemented
