+++ src/gc_pcsaft/eos/mod.rs
@@ -1,8 +1,7 @@
 use crate::association::Association;
 use crate::hard_sphere::HardSphere;
-use feos_core::joback::Joback;
 use feos_core::parameter::ParameterHetero;
-use feos_core::{EquationOfState, HelmholtzEnergy, IdealGasContribution, MolarWeight};
+use feos_core::{Components, HelmholtzEnergy, MolarWeight, Residual};
 use ndarray::Array1;
 use quantity::si::*;
 use std::f64::consts::FRAC_PI_6;
@@ -43,7 +42,6 @@ pub struct GcPcSaft {
     pub parameters: Arc<GcPcSaftEosParameters>,
     options: GcPcSaftOptions,
     contributions: Vec<Box<dyn HelmholtzEnergy>>,
-    joback: Joback,
 }
 
 impl GcPcSaft {
@@ -72,18 +70,14 @@ impl GcPcSaft {
             contributions.push(Box::new(Dipole::new(&parameters)))
         }
         Self {
-            parameters: parameters.clone(),
+            parameters,
             options,
             contributions,
-            joback: parameters.joback_records.clone().map_or_else(
-                || Joback::default(parameters.chemical_records.len()),
-                Joback::new,
-            ),
         }
     }
 }
 
-impl EquationOfState for GcPcSaft {
+impl Components for GcPcSaft {
     fn components(&self) -> usize {
         self.parameters.molarweight.len()
     }
@@ -94,7 +88,9 @@ impl EquationOfState for GcPcSaft {
             self.options,
         )
     }
+}
 
+impl Residual for GcPcSaft {
     fn compute_max_density(&self, moles: &Array1<f64>) -> f64 {
         let p = &self.parameters;
         let moles_segments: Array1<f64> = p.component_index.iter().map(|&i| moles[i]).collect();
@@ -102,12 +98,8 @@ impl EquationOfState for GcPcSaft {
             / (FRAC_PI_6 * &p.m * p.sigma.mapv(|v| v.powi(3)) * moles_segments).sum()
     }
 
-    fn residual(&self) -> &[Box<dyn HelmholtzEnergy>] {
+    fn contributions(&self) -> &[Box<dyn HelmholtzEnergy>] {
         &self.contributions
-    }
-
-    fn ideal_gas(&self) -> &dyn IdealGasContribution {
-        &self.joback
     }
 }
