+++ src/lib.rs
@@ -33,10 +33,11 @@
 
 #![warn(clippy::all)]
 #![allow(clippy::too_many_arguments)]
-#[cfg(feature = "dft")]
-mod dft;
-#[cfg(feature = "dft")]
-pub use dft::FunctionalVariant;
+#![allow(deprecated)]
+// #[cfg(feature = "dft")]
+// mod dft;
+// #[cfg(feature = "dft")]
+// pub use dft::FunctionalVariant;
 mod eos;
 // pub use eos::EosVariant;
