+++ src/pcsaft/dft/mod.rs
@@ -2,9 +2,8 @@ use super::PcSaftParameters;
 use crate::association::Association;
 use crate::hard_sphere::{FMTContribution, FMTVersion};
 use crate::pcsaft::eos::PcSaftOptions;
-use feos_core::joback::Joback;
 use feos_core::parameter::Parameter;
-use feos_core::{IdealGasContribution, MolarWeight};
+use feos_core::MolarWeight;
 use feos_dft::adsorption::FluidParameters;
 use feos_dft::solvation::PairPotential;
 use feos_dft::{FunctionalContribution, HelmholtzEnergyFunctional, MoleculeShape, DFT};
@@ -28,7 +27,6 @@ pub struct PcSaftFunctional {
     fmt_version: FMTVersion,
     options: PcSaftOptions,
     contributions: Vec<Box<dyn FunctionalContribution>>,
-    joback: Joback,
 }
 
 impl PcSaftFunctional {
@@ -87,17 +85,11 @@ impl PcSaftFunctional {
             }
         }
 
-        let joback = match &parameters.joback_records {
-            Some(joback_records) => Joback::new(joback_records.clone()),
-            None => Joback::default(parameters.m.len()),
-        };
-
         (Self {
             parameters,
             fmt_version,
             options: saft_options,
             contributions,
-            joback,
         })
         .into()
     }
@@ -120,10 +112,6 @@ impl HelmholtzEnergyFunctional for PcSaftFunctional {
 
     fn contributions(&self) -> &[Box<dyn FunctionalContribution>] {
         &self.contributions
-    }
-
-    fn ideal_gas(&self) -> &dyn IdealGasContribution {
-        &self.joback
     }
 
     fn molecule_shape(&self) -> MoleculeShape {
